requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0
stable-baselines3>=2.0.0
gym>=0.26.0
optuna>=3.2.0
backtesting>=0.3.0
ta>=0.10.2
requests>=2.31.0
beautifulsoup4>=4.12.0
transformers>=4.30.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
python-telegram-bot>=20.0
streamlit>=1.25.0
matplotlib>=3.7.0
plotly>=5.14.0
joblib>=1.2.0
pyyaml>=6.0
optuna>=3.2.0
